backend/requirements.txt
fastapi
uvicorn
transformers[torch]
torch
datasets
pandas
scikit-learn
trafilatura
beautifulsoup4
sentence-transformers
nltk
pydantic
python-multipart
spacy
openai
python-dotenv
requests
numpy
